+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/sources/EdgeSource.scala
@@ -27,7 +27,7 @@ import uk.co.gresearch.spark.dgraph.connector.encoder.EdgeEncoder
 import uk.co.gresearch.spark.dgraph.connector.executor.DgraphExecutorProvider
 import uk.co.gresearch.spark.dgraph.connector.model.EdgeTableModel
 import uk.co.gresearch.spark.dgraph.connector.partitioner.PartitionerProvider
-import uk.co.gresearch.spark.dgraph.connector.{ClusterStateProvider, SchemaProvider, TableProviderBase, TargetsConfigParser, TripleTable}
+import uk.co.gresearch.spark.dgraph.connector.{ChunkSizeOption, ClusterStateProvider, SchemaProvider, TableProviderBase, TargetsConfigParser, TripleTable}
 
 class EdgeSource() extends TableProviderBase
   with TargetsConfigParser with SchemaProvider
@@ -48,7 +48,8 @@ class EdgeSource() extends TableProviderBase
     val partitioner = getPartitioner(schema, clusterState, options)
     val encoder = EdgeEncoder(schema.predicateMap)
     val execution = DgraphExecutorProvider()
-    val model = EdgeTableModel(execution, encoder)
+    val chunkSize = getIntOption(ChunkSizeOption, options)
+    val model = EdgeTableModel(execution, encoder, chunkSize)
     new TripleTable(partitioner, model, clusterState.cid)
   }