+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/sources/TripleSource.scala
@@ -62,7 +62,8 @@ class TripleSource() extends TableProviderBase
       case None => TypedTripleEncoder(schema.predicateMap)
     }
     val execution = DgraphExecutorProvider()
-    val model = TripleTableModel(execution, encoder)
+    val chunkSize = getIntOption(ChunkSizeOption, options)
+    val model = TripleTableModel(execution, encoder, chunkSize)
     new TripleTable(partitioner, model, clusterState.cid)
   }