+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/TestPartition.scala
@@ -17,7 +17,6 @@
 
 package uk.co.gresearch.spark.dgraph.connector
 
-import io.dgraph.DgraphProto.TxnContext
 import org.scalatest.FunSpec
 import uk.co.gresearch.spark.dgraph.DgraphTestCluster
 import uk.co.gresearch.spark.dgraph.connector.encoder.TypedTripleEncoder
@@ -43,7 +42,7 @@ class TestPartition extends FunSpec with SchemaProvider with DgraphTestCluster {
         val partition = Partition(targets, Option(schema.predicates), None)
         val encoder = TypedTripleEncoder(schema.predicateMap)
         val execution = DgraphExecutorProvider()
-        val model = TripleTableModel(execution, encoder)
+        val model = TripleTableModel(execution, encoder, None)
         assert(model.modelPartition(partition).length === 44)
       }