+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   default number of predicates per partition of `1000` from before 0.3.0 ([issue #22](https://github.com/G-Research/spark-dgraph-connector/issues/22)).
 - The `PredicatePartitioner` combined with `UidRangePartitioner` is the default partitioner now,
   except for loading wide nodes, which still uses `PredicatePartitioner` as the default.
+- Add stream-like reading of partitions from Dgraph. Partitions are slip into smaller chunks.
+  Works with predicate partitioning only, so it cannot be combined with uid partitioning.
 
 ### Security
 - Moved Google Guava dependency version fix to 24.1.1-jre due to [known security vulnerability